Redirect to first child
When the new setting "islandora_compound_object_redirect_to_first" is
TRUE and the current URL points to a compound object with children,
then a redirect is generated to go to the first child.
This solves the problem that blocks on compound views get the wrong
context, because without the redirect the current object is the
compound but the data of the first child is displayed.