Scroll active into view if in scrollable container
Add javascript to scroll the active (selected) child of a compound
into view. This happens when:
- the active (selected) child is inside a scrollable container
- the active (selected) child is not fully visible inside that
  scrollable container or the next child is not (partly) visible
If both conditions are met, then the scrollable container is scrolled
in such a way that the active (selected) child is visible close to
the top or the left but (part of) the previous child is also visible.