Solve bug in IE
In IE the list did not scroll to the seleced item, because the
scrollWidth was 1 pixel larger than the clientWidth when a vertical
scroll bar was visible for the parent element. The loop was exited
prematurely due to the return false. Removing the return false
statement will make it work but have a small performance hit.
Also added some space above the selected item.